reinstate supported-targets; unit test to make sure it's correct
