+++ supported-targets
@@ -1,0 +1,1 @@
+x86_64-pc-windows-msvc x86_64-apple-darwin aarch64-apple-darwin x86_64-unknown-linux-gnu x86_64-unknown-linux-musl aarch64-unknown-linux-gnu aarch64-unknown-linux-musl aarch64-pc-windows-msvc armv7-unknown-linux-musleabihf armv7-unknown-linux-gnueabihf
